pyproject.toml
[tool.poetry]
name = "project-template"
version = "0.1.0"
description = ""
authors = ["Alexandr Shilin <alexandr-shilin@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django-rest-framework = "^0.1.0"
django-configurations = "^2.4.1"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
typing-extensions = "^4.8.0"
Pillow = "^10.0.1"
django-admin-rangefilter = "^0.11.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
black = {extras = ["d"], version = "^23.3.0"}
mypy = "<1.2.0"
pylint = "^2.17.2"
flake8 = "^6.0.0"
django-stubs = {extras = ["compatible-mypy"], version = "^1.16.0"}
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"
pylint-django = "^2.5.3"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
